birdtag-system/birdtag-system-6.6.3/src/handlers/requirements.txt
# AWS SDK
boto3==1.34.69
botocore==1.34.69

# Image processing
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

# Bird detection model
ultralytics==8.1.28
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
supervision==0.18.0

# Audio processing
librosa==0.10.1
tensorflow==2.15.1
soundfile==0.12.1
resampy==0.4.2
scikit-learn==1.6.1
tqdm==4.66.1

# Authentication
bcrypt==4.0.1
PyJWT==2.8.0
python-jose==3.3.0

# Utilities
requests==2.31.0
python-multipart==0.0.9